+++ src/main/java/calculator/domain/OperationFactory.java
@@ -1,4 +1,4 @@
-package calculator.service;
+package calculator.domain;
 
 import calculator.domain.*;
